+++ desktop/cmp/panel/impl/ResizeContainer.ts
@@ -8,6 +8,7 @@ import composeRefs from '@seznam/compose-react-refs';
 import {box, hbox, vbox} from '@xh/hoist/cmp/layout';
 import {hoistCmp, useContextModel} from '@xh/hoist/core';
 import {Children} from 'react';
+import {isString} from 'lodash';
 import {PanelModel} from '../PanelModel';
 import {dragger} from './dragger/Dragger';
 import {splitter} from './Splitter';
@@ -18,11 +19,15 @@ export const resizeContainer = hoistCmp.factory({
     className: 'xh-resizable',
 
     render({className, children}, ref) {
-        const panelModel = useContextModel(PanelModel);
-        let {size, resizable, collapsed, vertical, contentFirst, showSplitter} = panelModel,
+        const panelModel = useContextModel(PanelModel),
+            {size, resizable, collapsed, vertical, contentFirst, showSplitter} = panelModel,
             dim = vertical ? 'height' : 'width',
             child = Children.only(children),
-            items = [collapsed ? box(child) : box({item: child, [dim]: size})];
+            dragBarWidth = showSplitter ? '8px' : '0px',
+            sizeIsPct = isString(size) && size.endsWith('%');
+
+        const boxSize = sizeIsPct ? `calc(100% - ${dragBarWidth})` : size;
+        let items = [collapsed ? box(child) : box({item: child, [dim]: boxSize})];
 
         if (showSplitter) {
             const splitterCmp = splitter();
@@ -34,7 +39,9 @@ export const resizeContainer = hoistCmp.factory({
         }
 
         const cmp = vertical ? vbox : hbox,
-            maxDim = vertical ? 'maxHeight' : 'maxWidth';
+            maxDim = vertical ? 'maxHeight' : 'maxWidth',
+            minDim = vertical ? 'minHeight' : 'minWidth',
+            cmpSize = !collapsed && sizeIsPct ? size : undefined;
 
         if (panelModel._resizeRef) {
             ref = composeRefs(panelModel._resizeRef, ref);
@@ -44,7 +51,9 @@ export const resizeContainer = hoistCmp.factory({
             ref,
             className,
             flex: 'none',
+            [dim]: cmpSize,
             [maxDim]: '100%',
+            [minDim]: dragBarWidth,
             items
         });
     }