+++ desktop/cmp/panel/impl/dragger/DraggerModel.ts
@@ -55,7 +55,8 @@ export class DraggerModel extends HoistModel {
     private onDragStart = e => {
         const dragger = e.target;
         this.panelEl = dragger.parentElement;
-        const {panelEl: panel, panelModel} = this;
+        const {panelEl: panel, panelModel} = this,
+            {vertical} = panelModel;
 
         throwIf(
             !panel.nextElementSibling && !panel.previousElementSibling,
@@ -68,7 +69,7 @@ export class DraggerModel extends HoistModel {
 
         const {clientX, clientY} = this.parseEventPositions(e);
         this.resizeState = {startX: clientX, startY: clientY};
-        this.startSize = panelModel.size;
+        this.startSize = panel[vertical ? 'offsetHeight' : 'offsetWidth'];
         this.panelParent = panel.parentElement;
         panelModel.setIsResizing(true);